DatasetHero, extracted RawView, TableView cleanup to avoid double scroll bars
