+++ databrowser/src/components/download/DownloadSection.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="fixed bottom-0 left-0 py-5 w-full bg-white">
+  <div class="pt-6 lg:pt-7 w-full bg-white">
     <div class="mx-auto max-w-5xl">
       <div class="hidden lg:flex items-stretch space-x-4 h-full">
         <span class="self-center">{{ $t('datasets.download.info') }}</span>
