+++ databrowser/src/index.css
@@ -9,7 +9,6 @@ html {
 body {
   @apply text-gray-900;
   height: 100vh;
-  width: 100%;
   font-family: 'SourceSansPro', ui-sans-serif, system-ui, -apple-system,
     BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, 'Noto Sans',
     sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol',
@@ -17,5 +16,5 @@ body {
 }
 
 #app {
-  height: 100%;
+  height: 100vh;
 }
