+++ databrowser/src/pages/datasets/TableViewPage.vue
@@ -1,39 +1,24 @@
 <template>
-  <TableViewLayout>
-    <Hero>
-      <HeroTitle>All POIs</HeroTitle>
-      <HeroSubTitle
-        >Lorem ipsum dolor sit amen dolr sit amen dolr ist ist amen.
-      </HeroSubTitle>
-      <InternalLink to="/" tone="primary">More information</InternalLink>
-    </Hero>
-
-    <div class="flex flex-col h-full">
-      <DatasetNavigation :current-view="currentView" />
-      <TableView></TableView>
-    </div>
-  </TableViewLayout>
+  <AppLayout>
+    <DatasetHero />
+    <DatasetNavigation :current-view="currentView" />
+    <TableView></TableView>
+  </AppLayout>
 </template>
 
 <script lang="ts">
 import { defineComponent } from '@vue/runtime-core';
-import TableViewLayout from '../../layouts/TableViewLayout.vue';
-import Hero from '../../components/hero/HeroContainer.vue';
-import HeroTitle from '../../components/hero/HeroTitle.vue';
-import HeroSubTitle from '../../components/hero/HeroSubTitle.vue';
-import InternalLink from '../../components/link/InternalLink.vue';
 import TableView from '../../domain/datasets/tableView/TableView.vue';
 import { ViewPill } from '../../domain/datasets/navigation/types';
 import DatasetNavigation from '../../domain/datasets/navigation/DatasetNavigation.vue';
+import DatasetHero from '../../domain/datasets/DatasetHero.vue';
+import AppLayout from '../../layouts/AppLayout.vue';
 
 export default defineComponent({
   components: {
+    AppLayout,
+    DatasetHero,
     DatasetNavigation,
-    TableViewLayout,
-    Hero,
-    HeroTitle,
-    HeroSubTitle,
-    InternalLink,
     TableView,
   },
   setup() {
